requirements.txt
streamlit
streamlit_option_menu
firebase_admin
python-dotenv==0.21.1